Fixing logic to retrieve only personal clients transactions and products

+++ Clients/ClientController.cs
@@ -1,13 +1,15 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Myapp.GeneralClass;
 using Myapp.Models;
 using MyApp.GeneralClass;
+using MyApp.Products;
 
 namespace Myapp.Clients
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class ClientsController : ControllerBase
+    public class ClientsController : BaseController
     {
         private readonly ClientService _clientService;
 
@@ -17,7 +19,7 @@ namespace Myapp.Clients
         }
 
         //GET: api/clients
-        [Authorize(Roles = "user,admin")]
+        [Authorize(Roles = "admin")]
         [HttpGet]
         public async Task<ActionResult<ApiResponse<List<ClientDTO>>>> GetClients(){
             try{
@@ -41,18 +43,51 @@ namespace Myapp.Clients
             }
         }
 
+
+        //GET: api/clients
+        [Authorize(Roles = "user,admin")]
+        [HttpGet("my-clients")]
+        public async Task<ActionResult<ApiResponse<List<ClientDTO>>>> GetMyClients()
+        {
+            try
+            {
+                var userId = GetCurrentUserId();
+                var clients = await _clientService.GetMyClientsAsync(userId);
+                var clientDTOs = _clientService.MapToListDTOs(clients);
+
+                var response = new ApiResponse<List<ClientDTO>>(
+                    success: true,
+                    message: "Clients retrieved successfully.",
+                    data: clientDTOs
+                );
+                return Ok(response);
+            }
+            catch (Exception ex)
+            {
+                var errorResponse = new ApiResponse<List<Client>>(
+                    success: false,
+                    message: $"An error occurred: {ex.Message}",
+                    data: null
+                );
+                return StatusCode(StatusCodes.Status500InternalServerError, errorResponse);
+            }
+        }
+
+
         //GET: api/clients/{id}
         [Authorize(Roles = "user,admin")]
         [HttpGet("{id}")]
         public async Task<ActionResult<ApiResponse<ClientDTO>>> GetClient(Guid id)
         {
             try{
+
+                var userId = GetCurrentUserId();
                 var client = await _clientService.GetClientAsync(id);
-                if (client == null)
+                if (client == null || client.CreatedBy != userId)
                 {
                         var notFoundResponse = new ApiResponse<ClientDTO>(
                         success: false,
-                        message: "Client not found.",
+                        message: "Client not found or access denied.",
                         data: null
                     );
                     return BadRequest(notFoundResponse);
@@ -80,6 +115,7 @@ namespace Myapp.Clients
         [HttpPost]
         public async Task<ActionResult<Client>> CreateClient([FromBody] Client client)
         {
+            client.CreatedBy = GetCurrentUserId();
             await _clientService.CreateClientAsync(client);
             return CreatedAtAction(nameof(GetClient), new { id = client.Id }, client);
         }
@@ -98,6 +134,19 @@ namespace Myapp.Clients
                         data: null
                     );
                     return BadRequest(badRequestResponse);
+                }
+
+                var userId = GetCurrentUserId();
+
+                var existingClient = await _clientService.GetClientAsync(id);
+                if (existingClient == null || existingClient.CreatedBy != userId)
+                {
+                    var notFoundResponse = new ApiResponse<ProductDTO>(
+                         success: false,
+                         message: "Client not found or access denied.",
+                         data: null
+                    );
+                    return BadRequest(notFoundResponse);
                 }
 
                 await _clientService.UpdateClientAsync(id, client);
@@ -127,6 +176,18 @@ namespace Myapp.Clients
         {
             try
             {
+                var userId = GetCurrentUserId();
+
+                var existingClient = await _clientService.GetClientAsync(id);
+                if (existingClient == null || existingClient.CreatedBy != userId)
+                {
+                    var notFoundResponse = new ApiResponse<ProductDTO>(
+                         success: false,
+                         message: "Client not found or access denied.",
+                         data: null
+                    );
+                    return BadRequest(notFoundResponse);
+                }
                 // Mettre à jour partiellement le client
                 await _clientService.UpdateClientPartialAsync(id, updateClientDTO);
 
@@ -169,6 +230,19 @@ namespace Myapp.Clients
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteClient(Guid id)
         {
+            var userId = GetCurrentUserId();
+
+            var existingClient = await _clientService.GetClientAsync(id);
+            if (existingClient == null || existingClient.CreatedBy != userId)
+            {
+                var notFoundResponse = new ApiResponse<ProductDTO>(
+                     success: false,
+                     message: "Client not found or access denied.",
+                     data: null
+                );
+                return BadRequest(notFoundResponse);
+            }
+
             await _clientService.DeleteClientAsync(id);
             return NoContent();
         }
